添加测试，调整接口
- Group 服务层 dismiss 添加是否是小组管理员判断
- Group 分发层 /dismiss 添加 session 有效性验证
- 完成 Group 服务层 add、dismiss 的测试

+++ src/Service/Group.ts
@@ -28,8 +28,13 @@ export async function add(group: Readonly<Omit<Group, 'id'>>, session: Readonly<
         new ResponseBody<Pick<Group, 'id'>>(true, '', {id: groupId}));
 }
 
-export async function dismiss(group: Readonly<Pick<Group, 'id'>>): Promise<ServiceResponse<void>>
+export async function dismiss(group: Readonly<Pick<Group, 'id'>>, session: Readonly<Session>): Promise<ServiceResponse<void>>
 {
+    if (!(await GroupFunction.isGroupAdmin(group, session)))
+    {
+        return new ServiceResponse<void>(403, {},
+            new ResponseBody<void>(false, '解散失败：您不是小组的管理员'));
+    }
     await GroupTable.deleteById(group.id);
     return new ServiceResponse<void>(200, {},
         new ResponseBody<void>(true));
